task_service/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9.6
celery>=5.2.7
pika>=1.3.1
django-cors-headers>=3.13
# djangorestframework-authtoken>=3.14
drf-yasg>=1.21.5
djangorestframework-simplejwt>=5.2.2
